% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{cfs_dr}
\alias{cfs_dr}
\title{cfs_dr}
\usage{
cfs_dr(formula, y0, ref = "g1", data1, data2, link = "logit")
}
\arguments{
\item{formula}{a formula: y ~ x1+x2}

\item{y0}{a grid of threshold values}

\item{ref}{reference group. ref=g1 or g2.}

\item{data1}{data for group 1}

\item{data2}{data for group 2}

\item{link}{link function: logit or probit}
}
\value{
a series of (counterfactual) distributions
}
\description{
compute (counterfactual) distributions using distribution regression
}
