% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{de_dr.plot}
\alias{de_dr.plot}
\title{de_dr.plot}
\usage{
de_dr.plot(object, kind = "agg", qs)
}
\arguments{
\item{object}{decomposition components from de_dr}

\item{kind}{kinds of decomposition.It could be "agg", "composition", "structure".}

\item{qs}{quantiles to compute}
}
\description{
plot the reselts from de_dr
}
