% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{de_dr}
\alias{de_dr}
\title{dec_dr}
\usage{
de_dr(formula, y0, ref = "g1", data1, data2, link = "logit", qs)
}
\arguments{
\item{formula}{a formula: y ~ x1+x2}

\item{y0}{a grid of values}

\item{ref}{reference group. ref=g1 or g2.}

\item{data1}{data for group 1}

\item{data2}{data for group 2}

\item{link}{link function: logit or probit}

\item{qs}{quantiles to compute}
}
\value{
decomposition components
}
\description{
decompose quantile differences
}
