% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{dec_dr}
\alias{dec_dr}
\title{dec_dr}
\usage{
dec_dr(object)
}
\arguments{
\item{object}{results from qs_dr}
}
\value{
decomposition components of the quantile differences
}
\description{
decompose quantile differences
}
\keyword{internal}
