% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{sd_dr}
\alias{sd_dr}
\title{sd_dr}
\usage{
sd_dr(formula, y0, ref = "g1", data1, data2, link = "logit", qs, B)
}
\arguments{
\item{formula}{a formula: y ~ x1+x2}

\item{y0}{a grid of threshold values}

\item{ref}{reference group. ref=g1 or g2.}

\item{data1}{data for group 1}

\item{data2}{data for group 2}

\item{link}{link function: logit or probit}

\item{qs}{quantiles to compute}

\item{B}{the number of bootstrap iterations}
}
\value{
standard deviations for decomposition components
}
\description{
compute standard deviation for components
}
\examples{
data(five)
na=five[,1:6]
eu=five[,7:12]
colnames=c("Mkt.RF","SMB","HML","RMW", "CMA","SMALL.LoBM")
colnames(na)=colnames
colnames(eu)=colnames
formula=SMALL.LoBM~Mkt.RF+SMB+HML+RMW+CMA
y=na$SMALL.LoBM
y0=seq(quantile(y,probs = 0.05),quantile(y,probs = 0.95),length=100)
qs=seq(0.1,0.9,length.out = 10)
B=3
sd_dr(formula,y0,ref="g1",data1=na,data2=eu,link="logit",qs,B)
}
