% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{sd_obd}
\alias{sd_obd}
\title{sd_obd}
\usage{
sd_obd(formula, ref = "g1", data1, data2, B)
}
\arguments{
\item{formula}{a formula: y ~ x1+x2}

\item{ref}{reference group. ref=g1 or g2.}

\item{data1}{data for group 1}

\item{data2}{data for group 2}

\item{B}{the number of bootstrap iterations}
}
\value{
standard deviations of aggregate and detailed decomposition components
}
\description{
compute standard deviation of Oaxaca-Blinder decomposition
}
\examples{
data(five)
na=five[,1:6]
eu=five[,7:12]
colnames=c("Mkt.RF","SMB","HML","RMW", "CMA","SMALL.LoBM")
colnames(na)=colnames
colnames(eu)=colnames
formula=SMALL.LoBM~Mkt.RF+SMB+HML+RMW+CMA
B=10
sd=sd_obd(formula,ref="g1",data1=na,data2=eu,B=B)
}
