% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{dr}
\alias{dr}
\title{dr}
\usage{
dr(formula, y0, data, link = "logit")
}
\arguments{
\item{formula}{a formula: y ~ x1+x2}

\item{y0}{a grid of threshold values}

\item{data}{data}

\item{link}{link function: logit or probit}
}
\value{
a list of models for y0
}
\description{
distribution regression
}
\examples{
data(five)
na=five[,1:6]
colnames=c("Mkt.RF","SMB","HML","RMW", "CMA","SMALL.LoBM")
colnames(na)=colnames
formula=SMALL.LoBM~Mkt.RF+SMB+HML+RMW+CMA
y=na$SMALL.LoBM
y0=seq(quantile(y,probs = 0.05),quantile(y,probs = 0.95),length=100)
model=dr(formula,y0=y0,data=na,link="logit")
}
