% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{dr}
\alias{dr}
\title{dr}
\usage{
dr(formula, y0, data, link = "logit")
}
\arguments{
\item{formula}{a formula: y ~ x1+x2}

\item{y0}{a grid of values}

\item{data}{data}

\item{link}{link function}
}
\value{
a list of results
}
\description{
distribution regression
}
