% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{obd}
\alias{obd}
\title{obd}
\usage{
obd(formula, ref = "g1", data1, data2)
}
\arguments{
\item{formula}{a formula: y ~ x1+x2}

\item{ref}{reference group. ref=g1 or g2.}

\item{data1}{data for group 1}

\item{data2}{data for group 2}
}
\value{
aggregate and detailed decomposition components
}
\description{
do Oaxaca-Blinder decomposition
}
