% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{qs_dr}
\alias{qs_dr}
\title{qs_dr}
\usage{
qs_dr(object, y0, qs)
}
\arguments{
\item{object}{results from cfs_dr}

\item{y0}{a grid of values}

\item{qs}{quantiles to compute}
}
\value{
a series of quantiles
}
\description{
inverse the distribution function to obtain quantiles
}
